+++ tests/integration/query_processing/test_btree.rs
@@ -434,7 +434,7 @@ fn write_at(io: &impl IO, file: Arc<dyn File>, offset: usize, data: &[u8]) {
         // reference the buffer to keep alive for async io
         let _buf = _buf.clone();
     });
-    let result = file.pwrite(offset, buffer, completion).unwrap();
+    let result = file.pwrite(offset as u64, buffer, completion).unwrap();
     while !result.is_completed() {
         io.run_once().unwrap();
     }
